requirements.txt
miniaudio==1.10
mutagenpip==1.44.0
peewee==3.13.2